Restructured project to use a simple configuration via CLI or in the source files.
