--- README.md
@@ -1,11 +1,0 @@
-# Sphere placing algorithm
-
-### Requirements
-- CMake
-- nlohmann Json library for C++
-
-### Installation
-- Download folder and let CMake do the configuration, under Windows 10 / 11 and Visual Studio 2022 it should work out of the box (if nlohmann::json is available)
-
-### Usage
-- Run the executable with a config.json file in the same folder, such as the one found in the example folder
