--- cmake/FindSomeLib.cmake
@@ -1,1 +1,0 @@
-
