--- docs/CMakeLists.txt
@@ -1,5 +1,0 @@
-set(DOXYGEN_EXTRACT_ALL YES)
-set(DOXYGEN_BUILTIN_STL_SUPPORT YES)
-
-doxygen_add_docs(docs modern/lib.hpp "${CMAKE_CURRENT_SOURCE_DIR}/mainpage.md"
-                 WORKING_DIRECTORY "${PROJECT_SOURCE_DIR}/include")
