+++ src/main.h
@@ -1,0 +1,13 @@
+﻿// sphere_placer.h : Include file for standard system include files,
+// or project specific include files.
+
+#pragma once
+
+#include <iostream>
+#include <fstream>
+#include <random>
+#include "Placer.h"
+#include "PlacerConfig.h"
+#include "RNG/RNG.h"
+#include "file_interface.h"
+
